refactor(pipeline): absorb goal extractor (CHR-X-029)
Move the stage-2 goal extractor into the pipeline authority, repoint its live consumers and tests, and refresh generated module truth surfaces.
