--- aragora/goals/__init__.py
@@ -1,10 +1,0 @@
-"""
-Goal Extraction Module.
-
-Transforms organized ideas (Stage 1) into actionable goals and principles
-(Stage 2) using AI-powered synthesis.
-"""
-
-from .extractor import GoalExtractionConfig, GoalExtractor, GoalGraph, GoalNode
-
-__all__ = ["GoalExtractionConfig", "GoalExtractor", "GoalGraph", "GoalNode"]
