+++ aragora/module_tiers.yaml
@@ -25,10 +25,10 @@ thresholds:
 
 summary:
   core: 21
-  integrated: 91
+  integrated: 90
   experimental: 29
   deprecated: 3
-  total: 144
+  total: 143
 
 modules:
   # --- core (21) ---
@@ -74,36 +74,36 @@ modules:
     override_reason: "database abstraction"
   - name: debate
     tier: core
-    py_files: 265
-    importer_count: 160
-    test_file_count: 611
+    py_files: 267
+    importer_count: 162
+    test_file_count: 614
     override_reason: "Arena + DebateProtocol + consensus — mainline flow"
   - name: events
     tier: core
-    py_files: 26
+    py_files: 21
     importer_count: 129
-    test_file_count: 58
+    test_file_count: 59
   - name: gauntlet
     tier: core
     py_files: 39
     importer_count: 66
-    test_file_count: 112
+    test_file_count: 114
   - name: knowledge
     tier: core
     py_files: 178
     importer_count: 171
-    test_file_count: 236
+    test_file_count: 237
   - name: memory
     tier: core
-    py_files: 57
+    py_files: 58
     importer_count: 112
-    test_file_count: 119
+    test_file_count: 121
     override_reason: "ContinuumMemory + CritiqueStore — persistent state"
   - name: nomic
     tier: core
-    py_files: 162
-    importer_count: 89
-    test_file_count: 223
+    py_files: 165
+    importer_count: 88
+    test_file_count: 225
   - name: observability
     tier: core
     py_files: 92
@@ -111,9 +111,9 @@ modules:
     test_file_count: 57
   - name: pipeline
     tier: core
-    py_files: 50
-    importer_count: 72
-    test_file_count: 125
+    py_files: 51
+    importer_count: 73
+    test_file_count: 131
   - name: ranking
     tier: core
     py_files: 26
@@ -133,13 +133,13 @@ modules:
   - name: resilience
     tier: core
     py_files: 13
-    importer_count: 154
+    importer_count: 153
     test_file_count: 97
   - name: server
     tier: core
-    py_files: 1117
-    importer_count: 162
-    test_file_count: 1887
+    py_files: 1120
+    importer_count: 160
+    test_file_count: 1890
     override_reason: "FastAPI server + 3K+ API operations"
   - name: storage
     tier: core
@@ -148,7 +148,7 @@ modules:
     test_file_count: 166
     override_reason: "persistence layer"
 
-  # --- integrated (91) ---
+  # --- integrated (90) ---
   - name: analysis
     tier: integrated
     py_files: 29
@@ -238,7 +238,7 @@ modules:
     tier: integrated
     py_files: 44
     importer_count: 46
-    test_file_count: 57
+    test_file_count: 58
   - name: coordination
     tier: integrated
     py_files: 11
@@ -309,11 +309,6 @@ modules:
     py_files: 7
     importer_count: 10
     test_file_count: 12
-  - name: goals
-    tier: integrated
-    py_files: 2
-    importer_count: 8
-    test_file_count: 18
   - name: gti
     tier: integrated
     py_files: 5
@@ -434,7 +429,7 @@ modules:
     tier: integrated
     py_files: 27
     importer_count: 138
-    test_file_count: 26
+    test_file_count: 28
   - name: plugins
     tier: integrated
     py_files: 11
@@ -467,9 +462,9 @@ modules:
     test_file_count: 14
   - name: queue
     tier: integrated
-    py_files: 17
-    importer_count: 9
-    test_file_count: 18
+    py_files: 13
+    importer_count: 7
+    test_file_count: 17
   - name: ralph
     tier: integrated
     py_files: 7
@@ -489,7 +484,7 @@ modules:
     tier: integrated
     py_files: 14
     importer_count: 14
-    test_file_count: 29
+    test_file_count: 30
   - name: rlm
     tier: integrated
     py_files: 31
@@ -549,7 +544,7 @@ modules:
     tier: integrated
     py_files: 126
     importer_count: 32
-    test_file_count: 199
+    test_file_count: 201
   - name: templates
     tier: integrated
     py_files: 1
