+++ tests/canvas/test_idea_to_execution.py
@@ -36,7 +36,7 @@ from aragora.canvas.stages import (
     StageTransition,
     content_hash,
 )
-from aragora.goals.extractor import GoalExtractor, GoalGraph, GoalNode
+from aragora.pipeline.goal_extractor import GoalExtractor, GoalGraph, GoalNode
 from aragora.pipeline.idea_to_execution import IdeaToExecutionPipeline, PipelineResult
