+++ tests/goals/test_clustering.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 
 import pytest
 
-from aragora.goals.extractor import (
+from aragora.pipeline.goal_extractor import (
     GoalExtractor,
     _jaccard_similarity,
     _name_cluster,
