+++ tests/pipeline/test_adapters.py
@@ -21,7 +21,7 @@ from aragora.canvas.stages import (
     StageEdgeType,
     content_hash,
 )
-from aragora.goals.extractor import GoalNode
+from aragora.pipeline.goal_extractor import GoalNode
 from aragora.pipeline.adapters import (
     canvas_to_universal_graph,
     from_argument_node,
