+++ tests/pipeline/test_pipeline_e2e_flow.py
@@ -17,7 +17,7 @@ from aragora.pipeline.idea_to_execution import (
     StageResult,
 )
 from aragora.canvas.stages import PipelineStage, content_hash
-from aragora.goals.extractor import GoalGraph, GoalNode
+from aragora.pipeline.goal_extractor import GoalGraph, GoalNode
 
 
 # =============================================================================
