+++ tests/pipeline/test_pipeline_smoke.py
@@ -25,7 +25,7 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from aragora.canvas.stages import GoalNodeType, PipelineStage
-from aragora.goals.extractor import GoalExtractor, GoalGraph, GoalNode
+from aragora.pipeline.goal_extractor import GoalExtractor, GoalGraph, GoalNode
 from aragora.pipeline.idea_to_execution import (
     IdeaToExecutionPipeline,
     PipelineConfig,
